Phase C: asset categories linking accounts, richer fixed-asset fields
- AssetCategory model (group/category → asset, accumulated-depreciation,
  and expense accounts) with new asset-categories CRUD module; linking a
  category's asset account auto-flags Account.isFixedAssetAccount, and
  unlinking/deleting un-flags it when no longer referenced.
- FixedAsset gains categoryId, depreciationStartDate, plateNumber, and
  custodianEmployeeId (simple non-financial custody tracking, distinct
  from EmployeeAdvance); usefulLifeYears widened from Int to Decimal(6,2)
  to support fractional years.
- Fixed-assets service derives the posting account from the chosen
  category (falling back to a raw account for legacy/uncategorized
  assets); depreciation posting now prefers each asset's category-specific
  expense/accumulated-depreciation accounts, and disposal prefers the
  category's accumulated-depreciation account, both falling back to the
  shared flat accounts.
- New frontend "إعدادات الأصول الثابتة" settings tab, EmployeeSearchSelect
  component, and AssetRegisterTab updated to pick a category (or fall
  back to a raw account) plus the new fields.

+++ src/modules/purchaseInvoices/purchaseInvoices.schemas.ts
@@ -7,7 +7,7 @@ const lineSchema = z.object({
   itemId: z.string().optional(),
   warehouseId: z.string().optional(),
   // فقط لو الصنف من نوع "أصل ثابت"
-  usefulLifeYears: z.coerce.number().int().positive().optional(),
+  usefulLifeYears: z.coerce.number().positive().optional(),
   salvageValue: z.coerce.number().min(0).optional(),
   description: z.string().optional(),
   quantity: z.coerce.number().positive(),
